Remove unused option "include_pairs"
"include_pairs" is hard to compute.  How to deal with many attack cards
without defense cards is also a part of this game.  In addition, recent
expansions have no defense card.  It implies that new defense cards will
almost never be designed.

This option is not implemented and its UI is disabled.
So that it's safe to remove this option.

Close gh-48.